Refactored the Delegates and regex declarations
