--- Sources/Lux/InjectorImplementations/Json/JSONDelegate.swift
@@ -1,3 +1,0 @@
-import Foundation
-
-open class JSONDelegate: InjectorDelegate<JSONCategory> {}
