+++ Sources/Lux/InjectorImplementations/Json/JSONInjector.swift
@@ -1,5 +1,17 @@
 import Foundation
 
+// MARK: Regex
+
+extension RegexPattern {
+    static let json = RegexPattern(#"(\{|\}|\(|\)|\[|\]|,|"[^"]*"\s*:|"[^"]*")"#, type: .plain)
+}
+
+// MARK: Delegate
+
+public typealias JSONDelegate = InjectorDelegate<JSONCategory>
+
+// MARK: - Injector
+
 public final class JSONInjector<Output: Appendable, Injection: InjectionType, InjType: InjectorType<Output, Injection>>: BaseInjector<JSONCategory, Output, Injection, InjType> {
 
     // MARK: - Properties
