--- Sources/Lux/InjectorImplementations/Json/RegexPattern+Json.swift
@@ -1,3 +1,0 @@
-extension RegexPattern {
-    public static let json = RegexPattern(#"(\{|\}|\(|\)|\[|\]|,|"[^"]*"\s*:|"[^"]*")"#, type: .plain)
-}
