--- Sources/Lux/InjectorImplementations/Plist/PlistDelegate.swift
@@ -1,1 +1,0 @@
-open class PlistDelegate: InjectorDelegate<PlistCategory> {}
