+++ Sources/Lux/InjectorImplementations/Swift/SwiftDelegate+Extensions.swift
@@ -1,7 +1,7 @@
 import Foundation
 import Splash
 
-extension InjectorDelegate where Cat == SwiftCategory {
+extension SwiftDelegate {
 
     var theme: Theme { Theme(
     font: .init(size: 12),
@@ -19,5 +19,3 @@ extension InjectorDelegate where Cat == SwiftCategory {
     ]
     )}
 }
-
-open class SwiftDelegate: InjectorDelegate<SwiftCategory> {}
