+++ Sources/Lux/InjectorImplementations/XMLEnhanced/XMLEnhancedDelegate.swift
@@ -1,1 +1,1 @@
-open class XMLEnhancedDelegate: InjectorDelegate<XMLEnhancedCategory> {}
+
