--- Sources/Lux/InjectorImplementations/Xml/RegexPattern+Xml.swift
@@ -1,4 +1,0 @@
-public extension RegexPattern {
-    static let htmlXml = RegexPattern(#"&lt;("[^"]*"|'[^']*'|[^&]*)+&gt;"#, type: .html)
-    static let plainXml = RegexPattern(#"<("[^"]*"|'[^']*'|[^<^>])+>"#, type: .plain)
-}
