--- Sources/Lux/InjectorImplementations/Xml/XMLDelegate.swift
@@ -1,1 +1,0 @@
-open class XMLDelegate: InjectorDelegate<XMLCategory> {}
