+++ Sources/Lux/InjectorImplementations/Xml/XMLInjector.swift
@@ -1,5 +1,18 @@
 import Foundation
 
+// MARK: Regex
+
+extension RegexPattern {
+    static let htmlXml = RegexPattern(#"&lt;("[^"]*"|'[^']*'|[^&]*)+&gt;"#, type: .html)
+    static let plainXml = RegexPattern(#"<("[^"]*"|'[^']*'|[^<^>])+>"#, type: .plain)
+}
+
+// MARK: Delegate
+
+public typealias XMLDelegate = InjectorDelegate<XMLCategory>
+
+// MARK: - Injector
+
 /// Inject strings into a text depending on the configuration or the delegate.
 public final class XMLInjector<Output: Appendable, Injection: InjectionType, InjType: InjectorType<Output, Injection>>: BaseInjector<XMLCategory, Output, Injection, InjType> {
