--- Sources/Lux/InjectorImplementations/Zsh/ZshDelegate.swift
@@ -1,3 +1,0 @@
-import Foundation
-
-open class ZshDelegate: InjectorDelegate<ZshCategory> {}
