+++ Sources/Lux/InjectorImplementations/Zsh/ZshInjector.swift
@@ -1,11 +1,38 @@
 import Foundation
 
+// MARK: Regex
+
+extension RegexPattern {
+
+    static let zsh = RegexPattern(
+          #""[^"]*"|'[^']*'"# // strings
+        + #"|\s?\x5C#[^\s]*|#.*(?=\n|\Z)"# // comments and escaped # signs
+        + #"|(\s?sudo|\$\(|\$|\[|`|\n|\r|\|)\h*[a-zA-Z0-9]{1}[a-zA-Z0-9_-]*=?"# // programs and variables defs
+        + #"|\s\h*[a-zA-Z0-9\/\.]{1}[^\s]*"# // commands and option values
+        + #"|\s-{1,2}[a-zA-Z0-9_-]+"# // options and flags
+        + #"|\$\{[a-zA-Z0-9_-]+\}|"# // variable with brackets ${variable}
+        + #"\[|\]|;|\(|\)|\{|\}|`"#, // punctuation
+        type: .plain)
+
+    /// Find variables in a string in Zsh
+    static let zshVariables = RegexPattern(
+        #"\$[a-zA-Z0-9_-]+"#
+        + #"|\$\{[a-zA-Z0-9_-]+\}"#,
+        type: .plain)
+}
+
+// MARK: Delegate
+
+public typealias ZshDelegate = InjectorDelegate<ZshCategory>
+
+// MARK: - Injector
+
 public final class ZshInjector<Output: Appendable, Injection: InjectionType, InjType: InjectorType<Output, Injection>>: BaseInjector<ZshCategory, Output, Injection, InjType> {
 
     // MARK: - Properties
 
-    override var plainRegexPattern: RegexPattern { .plainZsh }
-    override var htmlRegexPattern: RegexPattern { .plainZsh }
+    override var plainRegexPattern: RegexPattern { .zsh }
+    override var htmlRegexPattern: RegexPattern { .zsh }
 
     // MARK: - Initialisation
